+++ release.clj
@@ -41,15 +41,16 @@
 
     (update-file "windows-amd64/project.clj" old->new)
     (update-file "ubuntu-latest-amd64/project.clj" old->new)
+    (update-file "ubuntu-latest-aarch64/project.clj" old->new)
     (update-file "macos-latest-amd64/project.clj" old->new)
     (update-file "macos-latest-aarch64/project.clj" old->new)
 
     (update-file "ubuntu-latest-amd64-shared/project.clj" old->new)
+    (update-file "ubuntu-latest-aarch64-shared/project.clj" old->new)
     (update-file "macos-latest-amd64-shared/project.clj" old->new)
     (update-file "macos-latest-aarch64-shared/project.clj" old->new)
 
     (update-file "x86_64-windows-gnu/project.clj" old->new)
-    (update-file "aarch64-linux-gnu/project.clj" old->new)
     ))
 
 (defn make-commit []
@@ -58,13 +59,14 @@
       "CHANGELOG.md"
       "windows-amd64/project.clj"
       "ubuntu-latest-amd64/project.clj"
+      "ubuntu-latest-aarch64/project.clj"
       "macos-latest-amd64/project.clj"
       "macos-latest-aarch64/project.clj"
 
       "x86_64-windows-gnu/project.clj"
-      "aarch64-linux-gnu/project.clj"
 
       "ubuntu-latest-amd64-shared/project.clj"
+      "ubuntu-latest-aarch64-shared/project.clj"
       "macos-latest-amd64-shared/project.clj"
       "macos-latest-aarch64-shared/project.clj")